refactor: move the transactional path off Pipeline onto its own Transaction type
`Pipeline` is defined by `|`: N stages running concurrently, stdout wired to
stdin through kernel pipes. `run_transactional` ran the same object
sequentially with no pipes and inherited stdio — two incompatible execution
models on one builder, selected by method name. `a | grep foo | c` under
`run_transactional()` handed `grep` the parent's stdin, silently and with no
error, and there was no way to reject it: `BitOr` is available on every
`Pipeline` and the `pub stages` field gives no signal to discriminate on.

Introduce `Transaction`, constructible only from an explicit stage list, with
a private field and deliberately no `BitOr`/`From<Pipeline>` impl. A
`|`-built chain can no longer reach the sequential runner at all, so the
footgun is closed by construction rather than by a doc comment.

This is a pure move: `TxnOutcome`, the guardrails and the coordinator are
carried over unchanged and the 15 existing tests keep passing unmodified
apart from their construction site. Nothing to break downstream —
`Pipeline::run_transactional` was never released, and is absent from the FFI,
the header, Python and Go.

Side effect worth noting: the `pipeline` module header's claim that "data
flows through kernel pipe buffers between stages" is true again for
everything left in that module.

Naming: `Transaction::new` here takes stages, not sandboxes. RFC #65's
end-state sketch is `Transaction::new(sandboxes).run()`, which cannot be
built as written — a `Sandbox` carries no argv, and the coordinator needs
the command. Happy to reserve `new` for that signature and rename this to
`Transaction::from_stages` if preferred.
