+++ crates/sandlock-core/src/lib.rs
@@ -24,6 +24,7 @@ pub mod netlink;
 pub(crate) mod procfs;
 pub(crate) mod port_remap;
 pub mod pipeline;
+pub mod transaction;
 pub mod policy_fn;
 pub mod image;
 pub mod fork;
@@ -40,7 +41,8 @@ pub use sandbox::{
     BindPorts, Confinement, ConfinementBuilder, Process, Sandbox, SandboxBuilder, StdioMode,
 };
 pub use result::{RunResult, ExitStatus};
-pub use pipeline::{Stage, Pipeline, Gather, TxnOutcome};
+pub use pipeline::{Stage, Pipeline, Gather};
+pub use transaction::{Transaction, TxnOutcome};
 pub use dry_run::{Change, ChangeKind, DryRunResult};
 // Sectioned-profile parsing types: ProfileInput is the top-level deserialization
 // target; ProgramSpec carries [program].exec/args (not a Sandbox field).
