+++ crates/sandlock-core/tests/integration.rs
@@ -31,8 +31,8 @@ mod test_landlock;
 #[path = "integration/test_pipeline.rs"]
 mod test_pipeline;
 
-#[path = "integration/test_pipeline_txn.rs"]
-mod test_pipeline_txn;
+#[path = "integration/test_transaction.rs"]
+mod test_transaction;
 
 #[path = "integration/test_network.rs"]
 mod test_network;
